update on values
for function of area of circle the on line 70 added a comma in between the radius and area. on line 72 changed the data type of the scanf for radius.

for the function of the rectangle on the line 119 we added "d" for the data type of the length. on line 125
changed the AREA to area to match the value